virtual_trainer_ai/backend/requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn[standard]
gunicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-multipart
bcrypt
PyJWT
redis
rq
boto3
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
structlog
